chore: ship tests/ via updater manifest + document tests/ layout and --only (#1440)
