+++ CONTRIBUTING.md
@@ -100,7 +100,22 @@ node cv-sync-check.mjs        # Config check
 # Dashboard
 npm run build:dashboard       # go build with platform-correct binary name
 npm run serve:dashboard       # launch the TUI against the repo root
+
+# Tests
+node test-all.mjs             # Full suite — run before pushing/opening a PR
+node test-all.mjs --quick     # Full suite, skipping the dashboard build
+node test-all.mjs --only providers/themuse   # Run just one provider's test(s)
 ```
+
+**Adding a test for a new scanner provider:** add one file at
+`tests/providers/{name}.test.mjs` — it's auto-discovered (`tests/**/*.test.mjs`),
+no registration needed. Do not add a section to `test-all.mjs` for this.
+
+**`--only` is a dev convenience, not a PR gate:** it runs *only* the discovered
+`tests/` files matching the given substring and skips every inline core
+section (syntax, scripts, dashboard, data contract, personal data, paths,
+etc.). A green `--only` run is **not** a green suite — always run the full
+`node test-all.mjs` before pushing.
 
 ## Brand and Trademark
